Refactor Oscillator Modules to MomentumOscillators for Clarity and Organization (MINOR) (#881)
