+++ src/main/java/com/verlumen/tradestream/backtesting/ParamConfigs.java
@@ -1,12 +1,12 @@
 package com.verlumen.tradestream.backtesting;
 
 import com.google.common.collect.ImmutableList;
-import com.verlumen.tradestream.backtesting.oscillators.OscillatorParams;
+import com.verlumen.tradestream.backtesting.momentumoscillators.MomentumOscillatorParams;
 
 final class ParamConfigs {
     static final ImmutableList<ParamConfig> ALL_CONFIGS = 
         ImmutableList.<ParamConfig>builder()
-            .addAll(OscillatorParams.ALL_CONFIGS)
+            .addAll(MomentumOscillatorParams.ALL_CONFIGS)
             .build();
 
     // Prevent instantiation
