+++ src/main/java/com/verlumen/tradestream/strategies/StrategyFactories.java
@@ -2,11 +2,11 @@ package com.verlumen.tradestream.strategies;
 
 import com.google.common.collect.ImmutableList;
 import com.verlumen.tradestream.strategies.movingaverages.MovingAverageStrategies;
-import com.verlumen.tradestream.strategies.oscillators.OscillatorStrategies;
+import com.verlumen.tradestream.strategies.momentumoscillators.OscillatorStrategies;
 
 /**
  * Provides a centralized collection of all available strategy factories across all categories.
- * This class aggregates factories from child packages (moving averages, oscillators, etc.)
+ * This class aggregates factories from child packages (moving averages, momentumoscillators, etc.)
  * and is immutable and thread-safe.
  */
 final class StrategyFactories {
