+++ src/main/java/com/verlumen/tradestream/strategies/momentumoscillators/BUILD
