+++ src/test/java/com/verlumen/tradestream/strategies/momentumoscillators/AdxStochasticStrategyFactoryTest.java
@@ -1,4 +1,4 @@
-package com.verlumen.tradestream.strategies.oscillators;
+package com.verlumen.tradestream.strategies.momentumoscillators;
 
 import static com.google.common.truth.Truth.assertThat;
